requirements.txt
faiss-cpu
sentence-transformers
tqdm
openai  
python-dotenv
google-generativeai
python-dotenv